+++ apps/ui-dev/component/actions/actiongroup/MainView.jsx
@@ -2,7 +2,7 @@ import ContentCopyFilled from "@elf-framework/icon/ContentCopyFilled";
 import DeleteFilled from "@elf-framework/icon/DeleteFilled";
 import EditFilled from "@elf-framework/icon/EditFilled";
 import MoreHorizOutlined from "@elf-framework/icon/MoreHorizOutlined";
-import { ActionGroup, Button, VBox } from "@elf-framework/ui";
+import { ActionGroup, Button, Tooltip, VBox } from "@elf-framework/ui";
 
 export function MainView() {
   return (
@@ -12,10 +12,12 @@ export function MainView() {
       <div>
         <div style={{ gap: 10, display: "flex", justifyContent: "center" }}>
           <ActionGroup>
-            <Button>
-              <EditFilled />
-              Edit
-            </Button>
+            <Tooltip message={"edit"}>
+              <Button>
+                <EditFilled />
+                Edit
+              </Button>
+            </Tooltip>
             <Button>
               <ContentCopyFilled />
               Copy
@@ -27,6 +29,14 @@ export function MainView() {
             <Button iconOnly>
               <MoreHorizOutlined />
             </Button>
+          </ActionGroup>
+          <ActionGroup>
+            <Tooltip message={"edit"}>
+              <Button>
+                <EditFilled />
+                Edit
+              </Button>
+            </Tooltip>
           </ActionGroup>
         </div>
       </div>