+++ packages/ui/src/components/property-editor/editor-items/LabelItem.jsx
@@ -1,0 +1,5 @@
+import { View } from "../../view";
+
+export function LabelItem({ style, item: { label } }) {
+  return <View style={style}>{label}</View>;
+}