+++ packages/ui/src/components/property-editor/editor-items/index.js
@@ -10,3 +10,4 @@ export * from "./SwitchItem";
 export * from "./TabContainerItem";
 export * from "./SliderItem";
 export * from "./DividerItem";
+export * from "./LabelItem";