+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @elf-framework/ui
 
+## 0.0.60
+
+### Patch Changes
+
+- add LabelItem to property editor
+
+- Updated dependencies []:
+  - @elf-framework/design-system@0.0.60
+
+## 0.0.58
+
+### Patch Changes
+
+- modify action-group button style bug
+
+- Updated dependencies []:
+  - @elf-framework/design-system@0.0.58
+
+## 0.0.57
+
+### Patch Changes
+
+- support tooltip for action group
+
+- Updated dependencies []:
+  - @elf-framework/design-system@0.0.57
+
+## 0.0.56
+
+### Patch Changes
+
+- update design system
+
 ## 0.0.54
 
 ### Patch Changes